chore: move changes from conf to custom_conf
